requirements.txt
playwright
python-dotenv
events